+++ ecco_v4_py/test/test_meridional_trsp.py
@@ -8,6 +8,7 @@ import pytest
 import ecco_v4_py
 
 from .test_common import llc_mds_datadirs, get_test_ds, get_test_vectors
+from .test_vector_calc import get_fake_vectors
 
 @pytest.mark.parametrize("lats",[-20,0,10,np.array([-30,-15,20,45])])
 def test_trsp_ds(get_test_ds,lats):
@@ -26,7 +27,7 @@ def test_vol_trsp(get_test_vectors,lats,basin):
     ds = get_test_vectors
     grid = ecco_v4_py.get_llc_grid(ds)
 
-    ds['U'],ds['V'] = get_fake_vectors(ds['U'].load(),ds['V'].load())
+    ds['U'],ds['V'] = get_fake_vectors(ds['U'],ds['V'])
     ds = ds.rename({'U':'UVELMASS','V':'VVELMASS'})
 
     trsp = ecco_v4_py.calc_meridional_vol_trsp(ds,lats,basin_name=basin,grid=grid)
@@ -56,7 +57,7 @@ def test_heat_trsp(get_test_vectors,lats,basin):
     ds = get_test_vectors
     grid = ecco_v4_py.get_llc_grid(ds)
 
-    ds['U'],ds['V'] = get_fake_vectors(ds['U'].load(),ds['V'].load())
+    ds['U'],ds['V'] = get_fake_vectors(ds['U'],ds['V'])
     ds = ds.rename({'U':'ADVx_TH','V':'ADVy_TH'})
     ds['DFxE_TH'] = ds['ADVx_TH'].copy()
     ds['DFyE_TH'] = ds['ADVy_TH'].copy()
@@ -88,7 +89,7 @@ def test_salt_trsp(get_test_vectors,lats,basin):
     ds = get_test_vectors
     grid = ecco_v4_py.get_llc_grid(ds)
 
-    ds['U'],ds['V'] = get_fake_vectors(ds['U'].load(),ds['V'].load())
+    ds['U'],ds['V'] = get_fake_vectors(ds['U'],ds['V'])
     ds = ds.rename({'U':'ADVx_SLT','V':'ADVy_SLT'})
     ds['DFxE_SLT'] = ds['ADVx_SLT'].copy()
     ds['DFyE_SLT'] = ds['ADVy_SLT'].copy()
@@ -111,19 +112,3 @@ def test_salt_trsp(get_test_vectors,lats,basin):
         test = trsp.sel(lat=lat).salt_trsp_z.reset_coords(drop=True)
         expected = (1e-6*(trspx+trspy)).reset_coords(drop=True)
         xr.testing.assert_allclose(test,expected)
-
-def get_fake_vectors(fldx,fldy):
-    for t in fldx.tile.values:
-        if t<6:
-            valx = 1.
-            valy = 1.
-        elif t == 6:
-            valx = 0.
-            valy = 0.
-        else:
-            valx = -1.
-            valy = 1.
-
-        fldx.loc[{'tile':t}] = valx
-        fldy.loc[{'tile':t}] = valy
-    return fldx, fldy
