+++ ecco_v4_py/test/test_stf.py
@@ -8,7 +8,7 @@ import pytest
 import ecco_v4_py
 
 from .test_common import llc_mds_datadirs, get_test_ds, get_test_vectors
-from .test_meridional_trsp import get_fake_vectors
+from .test_vector_calc import get_fake_vectors
 
 @pytest.mark.parametrize("lats",[-20,0,10,np.array([-30,-15,20,45])])
 @pytest.mark.parametrize("basin",[None,'atlExt','pacExt','indExt'])
@@ -19,7 +19,7 @@ def test_meridional_stf(get_test_vectors,lats,basin,doFlip):
     ds = get_test_vectors
     grid = ecco_v4_py.get_llc_grid(ds)
 
-    ds['U'],ds['V'] = get_fake_vectors(ds['U'].load(),ds['V'].load())
+    ds['U'],ds['V'] = get_fake_vectors(ds['U'],ds['V'])
     ds = ds.rename({'U':'UVELMASS','V':'VVELMASS'})
 
     trsp = ecco_v4_py.calc_meridional_stf(ds,lats,doFlip=doFlip,basin_name=basin,grid=grid)
@@ -64,7 +64,7 @@ def test_section_stf(get_test_vectors,name,pt1,pt2,maskW,maskS,expArr,doFlip):
     ds = get_test_vectors
     grid = ecco_v4_py.get_llc_grid(ds)
 
-    ds['U'],ds['V'] = get_fake_vectors(ds['U'].load(),ds['V'].load())
+    ds['U'],ds['V'] = get_fake_vectors(ds['U'],ds['V'])
     ds = ds.rename({'U':'UVELMASS','V':'VVELMASS'})
 
     if maskW is not None and maskS is not None:
